+++ crates/planner_app/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "planner_app"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+planning = { path = "../planning" }
+pnp = { path = "../pnp"}
+
+crux_core = { workspace = true }
+thiserror = { workspace = true }
+anyhow = { workspace = true }
+serde = { workspace = true }
+tracing = { workspace = true }
+crossbeam-channel = { workspace = true }
